Refactor localeUtils into its own file from modeUtils locale functions.
