+++ app/component/AddressRow.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 import StopCode from './StopCode';
-import { getTerminalOrStationText } from '../util/modeUtils';
+import { getTerminalOrStationText } from '../util/localeUtils';
 
 export default function AddressRow({
   desc,
